refactor test fixtures
